springboot-rest-api-security: adding springboot security
By simply adding the springbootsecurity dependency, you ll see if you hit your url now it would ask for the credentials before showing you the result. By default, username is 'user' and password would be printed in in the server logs. But this username and password can be overridded by mentioning it in the aaplication.properties file. And thats what I did here and override the username and password