springboot-rest-api-security: declarative vs programtic security
By default Spring provides two level of security i.e. authentication and authorization
Authentication: is this a valid user to the system or not.
Authorization: happens after the authentication and alright it is a valid user but does this user has the access to the requested resource or not.

there are two ways to define the security in spring i.e. declarative or programatic.
Declarative: They approach we use in the previous commit is a type of declarative approach in which we declare our users in application.properties file.  There is another way to achieve this objective. By defining a java class with 'Configuration' annotation (as in this commit). In this way you ll be able to manage the users more efficiently  and more naturally. This is also a way to use Declarative Security.

Programmatic Security: In most of the enterprise applications you may have to define the additional securities as per your business requirement on top of spring security. For that purpose we use the programmetic security which allows us the more flexibility to define custom security so that business needs can be full filled.

Note In later stages, I ll use the users and roles from the database. And if we get the chance, I ll shed some more light on progammetic security. so please make sure to go through the code commit by commit using 'cherry-pick' technique and must read the commit description to understand the concept completely.