springboot-rest-api-security: adding role base access
so far we have seem authentication . Even though I created the users with roles in last commit. But I haven't implemented their authorization to access. It mean, all those 3 users that I created in the last commit, had the same access to all the urls.

In this commit, I am adding the authorization to the url. For that I need to define a Bean in the SecurityConfig java class. And this method would be called upon before we entertain the the actual request just like the authentication screen. Here we are passing an object of HttpSecurity. And this object is helping us defining the role authorization using 'authorizeHttpRequest'. Here we are using lamba experssion to define these accesses. Here 'requestMatchers' method is taking two parameters i.e. Http request type (Get, Put, etc.) and the url on which we are defining the authorization. Next in the line is 'HasRole' method. it is use to declare which role has the access to this specific url. it only takes one role name param. to define mutliple role on same url, 'hasAnyRole' method use for this purpose.

We are using the basic Auth with simpleText password, So we need to define the auth type here as well.

Now about csrf disabling, by default, csrf is enabled to handle csrf forgery. but it is recommended to disable it if you are not using web browser for the request of type Put , Post  or Patch. If u don't disable the csrf here, it would expect a token named 'X-CSRF-TOKEN' in the header. And if you wont provide it in the request it would show you the 401 even with the valid  user and proper authorization. That why you need to disable it here.