05-springboot-rest-api-security:user and roles using JDBC
Earlier we were using the in-memory users and roles. It was more like hard coding approach. And I informed you guys that I ll move this to the JDBC in the later commits.  The purpose of using those in-memory users was keeping the things simple to get a good understanding of authentication and authorization.

Now that we have made ourselves familiar with the above mentioned concepts, the next thing is to use those users and roles from the databases. To achieve this, there are two possible ways. One Springboot provides us with out of the box support for user and roles. In which we don't have to right any additional code to manipulate data in database. But there is a constraint in this method and i.e. you have to use the same schema that is defined by the Springboot. I added a file for the required schema as per the springboot standard. It requires two tables i.e. users and authorities. Here authorities represents the user roles.

And the second approach is to use  the custom database tables . But in that case, we'll have to write some additional JDBC related code changes.

But for this commit, I am using the first approach. i.e. using spring defined schema for user and roles. Hence the very minimal code changes have done.