05-springboot-rest-api-security: saving hashed password instead of plaintext
from the start of the project we were saving our passwords as plaintext. But this is not a way to store password on production or any live application. to overcome this security concern, springboot provides us with the bcrypt hashing. Now to store the password in database, we need to write code to generate that hashing and save it in the database. but for now, I am using an online utility 'https://www.bcryptcalculator.com/'. So I  created the hashing using this utility, and save those passwords in database(see sql file added in this commit). the password is 'fun123' for the stated hashing in sql. And thats it. We don't have to do anything else . no code change because bcrypt has by default added in the springboot. so when it'll read {bcrypt} then it'll convert the input text in the bcrypt hash and compare it with the stored hash. if it matched then It means password is correct.

for further details on bcrypt, please go through the below mentioned link:
https://stackoverflow.com/questions/6832445/how-can-bcrypt-have-built-in-salts